requirements.txt
streamlit>=1.28.0
pyyaml>=6.0
python-dotenv>=1.0.0
gitpython>=3.1.40
pytest>=7.4.0
pytest-mock>=3.11.1

